backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3